Fix intermittent lockup in echo test
The NRF5x driver transmits a byte by writing it to the uart data
register and then waiting for the TXRDY event indicating that this
byte was sent. If another UART interrupt comes in at the right time
the the UART ISR handler will process and clear the TXRDY event,
even though this interrupt is not enabled. This causes serial_putc
to get stuck waiting on an already cleared TXRDY.

This patch fixing the lockup by preventing the UART ISR from handling
the TXRDY event if this interrupt is not enabled.